1.5.2 release. https://aws.amazon.com/releasenotes/PHP/1045686875113087

+++ lib/cachecore/cachefile.class.php
@@ -3,9 +3,10 @@
  * Container for all file-based cache methods. Inherits additional methods from <CacheCore>. Adheres
  * to the ICacheCore interface.
  *
- * @version 2009.10.10
- * @copyright 2006-2010 Ryan Parman
+ * @version 2012.01.28
+ * @copyright 2006-2012 Ryan Parman
  * @copyright 2006-2010 Foleeo, Inc.
+ * @copyright 2012 Amazon.com, Inc. or its affiliates.
  * @copyright 2008-2010 Contributors
  * @license http://opensource.org/licenses/bsd-license.php Simplified BSD License
  * @link http://github.com/skyzyx/cachecore CacheCore
@@ -44,12 +45,20 @@ class CacheFile extends CacheCore implements ICacheCore
 		{
 			return false;
 		}
-		elseif (file_exists($this->location) && is_writeable($this->location))
+		elseif (realpath($this->location) && file_exists($this->location) && is_writeable($this->location))
 		{
 			$data = serialize($data);
 			$data = $this->gzip ? gzcompress($data) : $data;
 
 			return (bool) file_put_contents($this->id, $data);
+		}
+		elseif (realpath($this->location) && file_exists($this->location))
+		{
+			throw new CacheFile_Exception('The file system location "' . $this->location . '" is not writable. Check the file system permissions for this directory.');
+		}
+		else
+		{
+			throw new CacheFile_Exception('The file system location "' . $this->location . '" does not exist. Create the directory, or double-check any relative paths that may have been set.');
 		}
 
 		return false;
@@ -100,6 +109,10 @@ class CacheFile extends CacheCore implements ICacheCore
 			$data = $this->gzip ? gzcompress($data) : $data;
 
 			return (bool) file_put_contents($this->id, $data);
+		}
+		else
+		{
+			throw new CacheFile_Exception('The file system location is not writeable. Check your file system permissions and ensure that the cache directory exists.');
 		}
 
 		return false;
@@ -168,3 +181,9 @@ class CacheFile extends CacheCore implements ICacheCore
 		return false;
 	}
 }
+
+
+/*%******************************************************************************************%*/
+// EXCEPTIONS
+
+class CacheFile_Exception extends CacheCore_Exception {}
